Bangalore Housing Prediction Using Historical Data & Publish the URL using AWS
By using the historical data of Banglore Housing Prices, created the website(URL) which predicts the Housing price for a given house, based upon the number of rooms, bathrooms, & location.
The usage of the model is strictly limited to the educational purpose, and the housing price is estimated entirely on the historical data.